Modify ReadME.md and script
